Move DateUtility unit tests (wrongly categorised as integration test) to phpunit
