+++ modules/tests/TestCaseList.php
@@ -5,7 +5,6 @@ class TestCaseList
         return array(
             array('DatabaseConnectionTest', 'DatabaseConnection Unit Tests'),
             array('DatabaseSearchTest',     'DatabaseSearch Unit Tests'),
-            array('DateUtilityTest',        'DateUtility Unit Tests'),
         );
     }
